+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Last Changes
 
+- [#9](https://github.com/LaxarApps/changelog-viewer-server/issues/9): cache: added support for refreshing caches
+
 
 ## v0.4.0